+++ interface99.h
@@ -40,6 +40,8 @@ SOFTWARE.
 #define declImpl       declImpl99
 #define externImpl     externImpl99
 #define externDeclImpl externDeclImpl99
+#define method         method99
+#define defaultMethod  defaultMethod99
 #define DYN            DYN99
 #define VTABLE         VTABLE99
 
@@ -62,6 +64,9 @@ SOFTWARE.
     ((iface){.self = (void *)(__VA_ARGS__), .vptr = &VTABLE99(implementer, iface)})
 #define VTABLE99(implementer, iface) ML99_CAT4(implementer, _, iface, _impl)
 
+#define method99(ret_ty, name, ...)        (0method, ret_ty, name, __VA_ARGS__)
+#define defaultMethod99(ret_ty, name, ...) (0defaultMethod, ret_ty, name, __VA_ARGS__)
+
 #define IFACE99_MAJOR 0
 #define IFACE99_MINOR 5
 #define IFACE99_PATCH 0
@@ -83,9 +88,9 @@ SOFTWARE.
  * // Only if <iface> is a marker interface without superinterfaces:
  * char dummy;
  *
- * <op-ret-ty>0 (*<op-name>0)(<op-params>0);
+ * <fn-ret-ty>0 (*<fn-name>0)(<fn-params>0);
  * ...
- * <op-ret-ty>N (*<op-name>N)(<op-params>N);
+ * <fn-ret-ty>N (*<fn-name>N)(<fn-params>N);
  *
  * const <requirement>0VTable *<requirement>;
  * ...
@@ -97,7 +102,7 @@ SOFTWARE.
         ML99_IF(                                                                                   \
             IFACE99_PRIV_IS_MARKER_IFACE(iface),                                                   \
             ML99_empty(),                                                                          \
-            v(iface##_INTERFACE(IFACE99_PRIV_OP_PTR, ~))),                                         \
+            IFACE99_PRIV_genFnPtrForEach(iface)),                                                  \
         ML99_IF(                                                                                   \
             IFACE99_PRIV_IS_SUB_IFACE(iface),                                                      \
             IFACE99_PRIV_genRequirementForEach(iface),                                             \
@@ -106,7 +111,15 @@ SOFTWARE.
 #define IFACE99_PRIV_genDummy(iface)                                                               \
     ML99_IF(IFACE99_PRIV_IS_EMPTY_VTABLE(iface), v(char dummy;), ML99_empty())
 
-#define IFACE99_PRIV_OP_PTR(_ctx, ret_ty, op_name, ...) ret_ty (*op_name)(__VA_ARGS__);
+/*
+ * <fn-ret-ty>0 (*<fn-name>0)(<fn-params>0);
+ * ...
+ * <fn-ret-ty>N (*<fn-name>N)(<fn-params>N);
+ */
+#define IFACE99_PRIV_genFnPtrForEach(iface)                                                        \
+    ML99_seqForEach(v(IFACE99_PRIV_genFnPtr), v(iface##_INTERFACE))
+
+#define IFACE99_PRIV_genFnPtr_IMPL(_tag, ret_ty, name, ...) v(ret_ty (*name)(__VA_ARGS__);)
 
 /*
  * const <requirement>0VTable *<requirement>;
@@ -140,9 +153,9 @@ SOFTWARE.
  * // Only if <iface> is a marker interface without superinterfaces:
  * .dummy = '\0',
  *
- * <op-name>0 = either <implementer>_<op-name>0 or <iface>_<op-name>0,
+ * <fn-name>0 = either <implementer>_<fn-name>0 or <iface>_<fn-name>0,
  * ...
- * <op-name>N = either <implementer>_<op-name>N or <iface>_<op-name>N,
+ * <fn-name>N = either <implementer>_<fn-name>N or <iface>_<fn-name>N,
  *
  * <requirement>0 = &VTABLE(<implementer, <requirement>0),
  * ...
@@ -154,7 +167,7 @@ SOFTWARE.
         ML99_IF(                                                                                   \
             IFACE99_PRIV_IS_MARKER_IFACE(iface),                                                   \
             ML99_empty(),                                                                          \
-            v(iface##_INTERFACE(IFACE99_PRIV_IMPL_OP, iface, implementer))),                       \
+            IFACE99_PRIV_genImplFnNameForEach(iface, implementer)),                                \
         ML99_IF(                                                                                   \
             IFACE99_PRIV_IS_SUB_IFACE(iface),                                                      \
             IFACE99_PRIV_genRequirementsImplForEach(iface, implementer),                           \
@@ -163,13 +176,23 @@ SOFTWARE.
 #define IFACE99_PRIV_genDummyInit(iface)                                                           \
     ML99_IF(IFACE99_PRIV_IS_EMPTY_VTABLE(iface), v(.dummy = '\0'), ML99_empty())
 
-#define IFACE99_PRIV_IMPL_OP(iface, implementer, _ret_ty, op_name, ...)                            \
-    .op_name = ML99_IF(                                                                            \
-        ML99_AND(                                                                                  \
-            IFACE99_PRIV_PROPERTY_PRESENT(iface##_##op_name##_DEFAULT),                            \
-            ML99_NOT(IFACE99_PRIV_PROPERTY_PRESENT(implementer##_##op_name##_CUSTOM))),            \
-        iface##_##op_name,                                                                         \
-        implementer##_##op_name),
+/*
+ * <fn-name>0 = either <implementer>_<fn-name>0 or <iface>_<fn-name>0,
+ * ...
+ * <fn-name>N = either <implementer>_<fn-name>N or <iface>_<fn-name>N,
+ */
+#define IFACE99_PRIV_genImplFnNameForEach(iface, implementer)                                      \
+    ML99_seqForEach(                                                                               \
+        ML99_appl(v(IFACE99_PRIV_genImplFnName), v(iface, implementer)),                           \
+        v(iface##_INTERFACE))
+
+#define IFACE99_PRIV_genImplFnName_IMPL(iface, implementer, tag, _ret_ty, fn_name, ...)            \
+    v(.fn_name = ML99_IF(                                                                          \
+          ML99_AND(                                                                                \
+              IFACE99_PRIV_IS_DEFAULT(tag),                                                        \
+              ML99_NOT(IFACE99_PRIV_IS_CUSTOM(implementer, fn_name))),                             \
+          iface##_##fn_name,                                                                       \
+          implementer##_##fn_name), )
 
 /*
  * <requirement>0 = &VTABLE(<implementer, <requirement>0),
@@ -195,18 +218,26 @@ SOFTWARE.
     const ML99_CAT(iface, VTable) VTABLE99(implementer, iface)
 // } (Implementation declaration)
 
+// Various predicates {
+
 #define IFACE99_PRIV_IS_EMPTY_VTABLE(iface)                                                        \
     ML99_AND(IFACE99_PRIV_IS_MARKER_IFACE(iface), ML99_NOT(IFACE99_PRIV_IS_SUB_IFACE(iface)))
-#define IFACE99_PRIV_IS_MARKER_IFACE(iface)                                                        \
-    ML99_VARIADICS_IS_SINGLE(iface##_INTERFACE(ML99_COMMA, ~))
-#define IFACE99_PRIV_IS_SUB_IFACE(iface) IFACE99_PRIV_PROPERTY_PRESENT(iface##_EXTENDS)
+#define IFACE99_PRIV_IS_MARKER_IFACE(iface) ML99_SEQ_IS_EMPTY(iface##_INTERFACE)
+#define IFACE99_PRIV_IS_SUB_IFACE(iface)    ML99_IS_TUPLE(iface##_EXTENDS)
 
-#define IFACE99_PRIV_PROPERTY_PRESENT(property) ML99_IS_TUPLE(property)
+#define IFACE99_PRIV_IS_DEFAULT(tag)           ML99_DETECT_IDENT(IFACE99_PRIV_IS_DEFAULT_, tag)
+#define IFACE99_PRIV_IS_DEFAULT_0defaultMethod ()
+
+#define IFACE99_PRIV_IS_CUSTOM(implementer, fn_name) ML99_IS_TUPLE(implementer##_##fn_name##_CUSTOM)
+// } (Various predicates)
 
 // Arity specifiers {
 
 #define IFACE99_PRIV_genRequirement_ARITY     1
 #define IFACE99_PRIV_genRequirementImpl_ARITY 2
+
+#define IFACE99_PRIV_genFnPtr_ARITY      1
+#define IFACE99_PRIV_genImplFnName_ARITY 2
 
 // Public: